Add compiler for fusion networks with LSG
Adds a function that compiles a fusion network with a partial order into
a sequence of instructions to be executed on our proposed architecture
with a single linear resource state generator and unbounded number of
measurements can happen in parallel.

Added tests which compile the instructions back into a measurement
pattern and confirm that it implements the fusion network and respects
the partial order.