Make inputs and outputs of OpenGraph ordered
To implement sequential composition of OpenGraphs, the inputs and
outputs need to be ordered, so we have stored them as lists now rather
than unordered sets.

Put measurements/inputs/outputs in networkx

By storing the measurement, input node and output node information
directory in the networkx graph as tags, we can better perform
isomorphism tests and it will make it easier to perform sequential and
parallel composition of graphs in future.